newly created data will now persist
Note: using any method to startup the docker containers ALSO re-executes anything within DockerFiles. The reason data didn't persist was because prisma:seed empties the database before seeding it. Added a new env variable and if statement within backend's DockerFile to combat this.
I don't know how to set this variable to true from the terminal, so it has to be set to true manually from the .env file for it to acivate the seed command again. (also discovered you can go into the node docker container's Exec tab in Docker Desktop and type the seed command there after starting the containers.)